+++ aegisub/src/frame_main_events.cpp
@@ -262,7 +262,7 @@ void FrameMain::RebuildRecentList(wxString listName,wxMenu *menu,int startID) {
 	wxString n;
 	wxArrayString entries = Options.GetRecentList(listName);
 	for (size_t i=0;i<entries.Count();i++) {
-		n = wxString::Format(_T("%i"),i+1);
+		n = wxString::Format(_T("%i"),(int)i+1);
 		if (i < 9) n = _T("&") + n;
 		wxFileName shortname(entries[i]);
 		wxString filename = shortname.GetFullName();
